#27 - Part the files in folders - Finish Page
